refactor: separate tw config colors for emails
