+++ examples/nextjs-comments-emails/tailwind.config.ts
@@ -1,23 +1,5 @@
-import type {
-  RecursiveKeyValuePair,
-  ResolvableTo,
-} from "tailwindcss/types/config";
 import type { Config } from "tailwindcss";
-
-// Export config specific to emails and define them to make them
-// usable in your code editor.
-export const emailColors: RecursiveKeyValuePair = {
-  foreground: "#111111",
-  mention: "#1667FF",
-  headline: "#171717",
-  comment: {
-    background: "#EEEEEE",
-    foreground: {
-      DEFAULT: "#171717",
-      subtle: "rgba(23, 23, 23, 0.6)",
-    },
-  },
-};
+import { emailColors } from "./emails/_tailwind/colors";
 
 const config = {
   content: [
